Move loader component to special modul
To work on it later
